first pass on fixing transplanted components
